fix: make importDataSet a single undoable action
