+++ v3/src/components/app.tsx
@@ -1,18 +1,16 @@
 import { observer } from "mobx-react-lite"
 import React, { useCallback, useEffect } from "react"
-import { kCaseTableTileType } from "./case-table/case-table-defs"
 import { CodapDndContext } from "./codap-dnd-context"
 import { Container } from "./container/container"
-import { createTileOfType } from "./create-tile"
 import { ToolShelf } from "./tool-shelf/tool-shelf"
 import { importV2Document } from "./import-v2-document"
 import { MenuBar } from "./menu-bar/menu-bar"
 import { appState } from "../models/app-state"
 import { addDefaultComponents } from "../models/codap/add-default-content"
 import {gDataBroker} from "../models/data/data-broker"
-import {DataSet, IDataSet, toCanonical} from "../models/data/data-set"
+import {IDataSet} from "../models/data/data-set"
 import { IDocumentModelSnapshot } from "../models/document/document"
-import { linkTileToDataSet } from "../models/shared/shared-data-utils"
+import { IImportDataSetOptions } from "../models/document/document-content"
 import { getSharedModelManager } from "../models/tiles/tile-environment"
 import { DocumentContext } from "../hooks/use-document-context"
 import {useDropHandler} from "../hooks/use-drop-handler"
@@ -20,7 +18,6 @@ import { useKeyStates } from "../hooks/use-key-states"
 import { registerTileTypes } from "../register-tile-types"
 import { importSample, sampleData } from "../sample-data"
 import { urlParams } from "../utilities/url-params"
-import t from "../utilities/translation/translate"
 
 import "../models/shared/shared-case-metadata-registration"
 import "../models/shared/shared-data-set-registration"
@@ -29,26 +26,13 @@ import "./app.scss"
 
 registerTileTypes([])
 
-interface IImportDataSetOptions {
-  createTableTile?: boolean // default true
-}
-export function handleImportDataSet(data: IDataSet, options?: IImportDataSetOptions) {
-  // add data set
-  const { sharedData } = gDataBroker.addDataSet(data)
-  if (sharedData.dataSet && (options?.createTableTile !== false)) {
-    // create the corresponding case table
-    const newTile = createTileOfType(kCaseTableTileType, appState.document.content)
-    if (newTile) {
-      // link the case table to the new data set
-      linkTileToDataSet(newTile.content, sharedData.dataSet)
-    }
-  }
-}
-
 export const App = observer(function App() {
-  const codapDocument = appState.document
-
   useKeyStates()
+
+  const handleImportDataSet = useCallback(
+    function handleImportDataSet(data: IDataSet, options?: IImportDataSetOptions) {
+      appState.document.content?.importDataSet(data, options)
+    }, [])
 
   const handleImportV3Document = useCallback((document: IDocumentModelSnapshot) => {
     appState.setDocument(document)
@@ -60,15 +44,6 @@ export const App = observer(function App() {
     onImportV2Document: importV2Document,
     onImportV3Document: handleImportV3Document
   })
-
-  function createNewStarterDataset() {
-    const attributeName = t("DG.AppController.createDataSet.initialAttribute")
-    const newData = [{[attributeName]: ""}]
-    const ds = DataSet.create({ name: t("DG.AppController.createDataSet.name")})
-    ds.addAttribute({ name: attributeName })
-    ds.addCases(toCanonical(ds, newData))
-    gDataBroker.addDataSet(ds)
-  }
 
   useEffect(() => {
     // connect the data broker to the shared model manager
@@ -86,14 +61,14 @@ export const App = observer(function App() {
           try {
             const data = await importSample(sample)
             // show case table if not showing a complete dashboard
-            handleImportDataSet(data, { createTableTile: !isDashboard })
+            appState.document.content?.importDataSet(data, { createDefaultTile: !isDashboard })
           }
           catch (e) {
             console.warn(`Failed to import sample "${sample}"`)
           }
         }
         else if (isDashboard) {
-          createNewStarterDataset()
+          appState.document.content?.createStarterDataset()
         }
         // we have to create a new starter data set only if none is imported to show the dashboard
         if (isDashboard) {
@@ -105,13 +80,14 @@ export const App = observer(function App() {
     initialize()
   }, [])
 
+  const documentContent = appState.document.content
   return (
     <CodapDndContext>
       <DocumentContext.Provider value={appState.document.content}>
         <div className="app" data-testid="app">
           <MenuBar/>
-          <ToolShelf content={codapDocument.content} createTileOfType={createTileOfType}/>
-          <Container content={codapDocument.content}/>
+          <ToolShelf content={documentContent}/>
+          <Container content={documentContent}/>
         </div>
       </DocumentContext.Provider>
     </CodapDndContext>
