+++ v3/src/test/test-tile-content.ts
@@ -1,6 +1,7 @@
 import { Instance } from "mobx-state-tree"
 import { TileContentModel } from "../models/tiles/tile-content"
 import { registerTileContentInfo } from "../models/tiles/tile-content-info"
+import { registerTileComponentInfo } from "../models/tiles/tile-component-info"
 
 export const TestTileContent = TileContentModel
   .named("TestTileContent")
@@ -14,4 +15,13 @@ registerTileContentInfo({
   prefix: "TEST",
   modelClass: TestTileContent,
   defaultContent: () => TestTileContent.create()
+})
+
+registerTileComponentInfo({
+  type: "Test",
+  TitleBar: () => null,
+  Component: () => null,
+  tileEltClass: "test-tile",
+  defaultWidth: 400,
+  defaultHeight: 300
 })
